// packages/core/src/modals/types/tools.d.ts
export type Tools =
    "bold"
    | "emoji"
    | "blockquote"
    | "code"
    | "codeBlock"
    | "heading"
    | "highlight"
    | "italic"
    | "link"
    | "orderList"
    | "taskList"
    | "subscript"
    | "superscript"
    | "table"
    | "textAlign"
    | "underline"
    | "strike"
    | "bulletList"
    | "orderedList"
    | "textStyle"
    | "undo"
    | "redo"